feat: Migrate context and func signatures, Misc
 - Create SetupSTGContext to be run in another Setup_____Context
   function
 - Migrate STGShur14Context, CreateSTGContext, and SetupSTGContext
   signatures to navierstokes.h
 - Add STG contexts to Physics and CeedData
 - Add missing CHKERRQ to PetscFClose
 - Move to SPDX license headers
